refactor: 할인 로직을 discount 도메인으로 분리
