--- src/refactoring/cart/models/product.ts
@@ -1,5 +1,0 @@
-export const getMaxDiscount = (
-  discounts: { quantity: number; rate: number }[]
-) => {
-  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
-};
